requirements.txt
PySide6
requests
cryptography
mnemonic
bip44
web3
qrcode
pillow